Add C++ Solution to Patterns/Math

+++ LeetCode/Patterns/Math/9. Palindrome Number/Solution.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         int temp = x;
-        int rev = 0;
+        long int rev = 0;
         while (x > 0){
             rev = (rev*10) + x%10;
             x = x/10;
